// frontend/src/api/models/frame.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { GroupFilter } from './groupFilter';
import type { SortKey } from './sortKey';

export interface Frame {
  /** @maxItems 10 */
  groupFilters: GroupFilter[];
  /** @maxItems 3 */
  sortKeys?: SortKey[];
  workflowId: string;
}